+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.XX]-2025-02-13
+
+### Added
+
+- Reference Values api endpoint:
+  -reference value by category
+
 ## [1.0.0-beta.1]-2025-02-11
 
 ### Added